Return image as file

+++ ViewModels/CandleItemBasicVM.cs
@@ -8,7 +8,7 @@
 		public string? Desciption { get; set; }
 		public string Category { get; set; } = null!;
 		public decimal Price { get; set; }
-		public int Height { get; set; }
+		public float Height { get; set; }
 		public int? BurningTime { get; set; }
 
 	}
